WEB-4526 move patientdata-related utils into dataViewUtils.js
