+++ app/pages/dashboard/PatientDrawer/useAgpCGM/getOpts.js
@@ -3,6 +3,7 @@ import _ from 'lodash';
 import get from 'lodash/get';
 import { utils as vizUtils } from '@tidepool/viz';
 import utils from '../../../../core/utils';
+import { getMostRecentDatumTimeByChartType } from '../../../../core/dataViewUtils';
 
 const getTimezoneFromTimePrefs = vizUtils.datetime.getTimezoneFromTimePrefs;
 
@@ -12,7 +13,7 @@ const getOpts = (
 ) => {
   const latestDatumByType = _.get(data, 'metaData.latestDatumByType');
   const mostRecentDatumDates = {
-    agpCGM: utils.getMostRecentDatumTimeByChartType(latestDatumByType, 'agpCGM'),
+    agpCGM: getMostRecentDatumTimeByChartType(latestDatumByType, 'agpCGM'),
   };
 
   const timePrefs = (() => {
